+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.11] - 2025-04-16
+
+- Update default thresholds for custom evals to 0.0 in `Validator` API.
+
+## [1.0.10] - 2025-04-15
+
+- Add async support to `Validator` API.
+
+## [1.0.9] - 2025-04-10
+
+- Refactor threshold validation in the `Validator` class to only check user-provided metrics.
+
 ## [1.0.8] - 2025-04-03
 
 - Update `Project.query()` method with optional `metadata` property to log and store arbitrary metadata.
@@ -47,7 +59,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release of the `cleanlab-codex` client library.
 
-[Unreleased]: https://github.com/cleanlab/cleanlab-codex/compare/v1.0.8...HEAD
+[Unreleased]: https://github.com/cleanlab/cleanlab-codex/compare/v1.0.11...HEAD
+[1.0.11]: https://github.com/cleanlab/cleanlab-codex/compare/v1.0.10...v1.0.11
+[1.0.10]: https://github.com/cleanlab/cleanlab-codex/compare/v1.0.9...v1.0.10
+[1.0.9]: https://github.com/cleanlab/cleanlab-codex/compare/v1.0.8...v1.0.9
 [1.0.8]: https://github.com/cleanlab/cleanlab-codex/compare/v1.0.7...v1.0.8
 [1.0.7]: https://github.com/cleanlab/cleanlab-codex/compare/v1.0.6...v1.0.7
 [1.0.6]: https://github.com/cleanlab/cleanlab-codex/compare/v1.0.5...v1.0.6